chore: Move old legacy stuff to a single source file
